支持高精度计算

+++ src/main/java/com/kamijoucen/ruler/component/IntegerNumberCache.java
@@ -2,8 +2,10 @@ package com.kamijoucen.ruler.component;
 
 import com.kamijoucen.ruler.domain.value.IntegerValue;
 
+import java.math.BigInteger;
+
 public interface IntegerNumberCache {
 
-    IntegerValue getValue(long num);
+    IntegerValue getValue(BigInteger num);
 
 }
